+++ services/workflows-service/scripts/workflows/runtime/generate-initial-collection-flow-example.ts
@@ -1,0 +1,59 @@
+import { PrismaClient } from '@prisma/client';
+
+export const generateInitialCollectionFlowExample = async (
+  prismaClient: PrismaClient,
+  {
+    workflowDefinitionId,
+    projectId,
+    endUserId,
+    businessId,
+    token,
+  }: {
+    workflowDefinitionId: string;
+    projectId: string;
+    endUserId: string;
+    businessId: string;
+    token: string;
+  },
+) => {
+  const creationArgs = {
+    data: {
+      endUserId: endUserId,
+      workflowDefinitionId: workflowDefinitionId,
+      projectId: projectId,
+      state: 'idle',
+      context: {
+        workflowId: workflowDefinitionId,
+        entity: {
+          ballerineEntityId: businessId,
+          type: 'business',
+          data: {
+            additionalInfo: {
+              mainRepresentative: {
+                firstName: 'John',
+                lastName: 'Doe',
+                email: 'test@gmail.com',
+              },
+            },
+          },
+        },
+        documents: [],
+      },
+      businessId: businessId,
+      workflowDefinitionVersion: 1,
+    },
+  };
+  const workflowRuntime = await prismaClient.workflowRuntimeData.create(creationArgs);
+
+  const workflowToken = await prismaClient.workflowRuntimeDataToken.create({
+    data: {
+      endUserId: endUserId,
+      token: token,
+      workflowRuntimeDataId: workflowRuntime.id,
+      projectId: projectId,
+      expiresAt: new Date(Date.now() + 30 * 24 * 60 * 60 * 1000),
+    },
+  });
+
+  return workflowToken;
+};