+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/headless-example
 
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.37
+  - @ballerine/workflow-browser-sdk@0.5.36
+
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.36
+  - @ballerine/workflow-browser-sdk@0.5.35
+
 ## 0.1.34
 
 ### Patch Changes