+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.36
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.36
+
+## 0.5.35
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.35
+
 ## 0.5.34
 
 ### Patch Changes